Update find.html
Deleted unneeded columns in the resultant csv file in order to more closely match resultant csv file to the original csv file. Now, the columns for each csv file are identical, but the resultant csv file additionally contains a "New Location" column of data in column E.

Also: CORS limitation can be bypassed by accessing the index.html and submitting your csv file while accessing index.html as a local file, and not as a file hosted on a web server such as github.io.